pyproject.toml
[project]
name = "backend-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agents>=1.4.0",
    "fastapi>=0.121.0",
    "openai>=2.7.1",
    "openai-agents>=0.5.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "uvicorn[standard]>=0.38.0",
]
